Reorganize test files by feature instead of origin
Dissolve design-doc-examples.test.ts, coverage-gaps.test.ts,
emitter.test.ts, and diagnostics.test.ts — distribute their tests
into feature-focused files. Create new files for unions, enums,
arrays, deprecation, circular references, and doc comments.

197 tests across 17 files (was 197 across 15).
