+++ packages/graphql/test/operation-kind.test.ts
@@ -1,7 +1,8 @@
+import { strictEqual } from "node:assert";
 import { expectDiagnostics, t } from "@typespec/compiler/testing";
 import { describe, expect, it } from "vitest";
 import { getOperationKind } from "../src/lib/operation-kind.js";
-import { Tester } from "./test-host.js";
+import { emitSingleSchema, Tester } from "./test-host.js";
 
 describe("Operation kinds", () => {
   it("declares a Mutation", async () => {
@@ -41,5 +42,53 @@ describe("Operation kinds", () => {
     ]);
     const operationKind = getOperationKind(program, testOperation);
     expect(operationKind).toBeUndefined();
+  });
+
+  describe("SDL output", () => {
+    it("emits queries, mutations, and subscriptions", async () => {
+      const code = `
+        @schema
+        namespace TestNamespace {
+          model User {
+            id: string;
+            name: string;
+          }
+
+          model Message {
+            id: string;
+            text: string;
+            userId: string;
+          }
+
+          @query
+          op getUser(id: string): User;
+
+          @query
+          op listUsers(): User[];
+
+          @mutation
+          op createUser(name: string): User;
+
+          @mutation
+          op updateUser(id: string, name: string): User;
+
+          @subscription
+          op onMessageReceived(userId: string): Message;
+        }
+      `;
+
+      const result = await emitSingleSchema(code, {});
+
+      strictEqual(result.includes("type Query {"), true);
+      strictEqual(result.includes("getUser(id: String!): User"), true);
+      strictEqual(result.includes("listUsers: [User!]"), true);
+
+      strictEqual(result.includes("type Mutation {"), true);
+      strictEqual(result.includes("createUser(name: String!): User"), true);
+      strictEqual(result.includes("updateUser(id: String!, name: String!): User"), true);
+
+      strictEqual(result.includes("type Subscription {"), true);
+      strictEqual(result.includes("onMessageReceived(userId: String!): Message"), true);
+    });
   });
 });
