+++ packages/graphql/test/schema.test.ts
@@ -1,7 +1,7 @@
 import { t } from "@typespec/compiler/testing";
 import { describe, expect, it } from "vitest";
 import { getSchema } from "../src/lib/schema.js";
-import { Tester } from "./test-host.js";
+import { emitWithDiagnostics, Tester } from "./test-host.js";
 
 describe("@schema", () => {
   it("Creates a schema with no name", async () => {
@@ -26,5 +26,35 @@ describe("@schema", () => {
 
     expect(schema?.type).toBe(TestNamespace);
     expect(schema?.name).toBe("MySchema");
+  });
+
+  it("emits separate schemas for different namespaces", async () => {
+    const code = `
+      @schema
+      namespace SchemaA {
+        model Foo {
+          id: string;
+        }
+
+        @query
+        op getFoo(): Foo;
+      }
+
+      @schema
+      namespace SchemaB {
+        model Bar {
+          name: string;
+        }
+
+        @query
+        op getBar(): Bar;
+      }
+    `;
+
+    const results = await emitWithDiagnostics(code, {});
+    // Should produce results (at least one schema)
+    expect(results.length).toBeGreaterThanOrEqual(1);
+    // First result should have output
+    expect(results[0].graphQLOutput).toBeDefined();
   });
 });
